Housecleaning (#270)
* - remove unused functions
- remove ruamel_yaml import from active scripts
- fix dontcare2background related to PR #256

* remove data_analysis.py from GDL and move to internal gitlab.
